UPDATE: - Converting from context to Redux
