+++ client/src/components/web-extension/CreateWallet/RecoveryPhrasePage.jsx
@@ -3,14 +3,12 @@ import { Button } from 'react-bootstrap';
 import { useSelector, useDispatch } from 'react-redux';
 import drop from 'lodash-es/drop';
 import dropRight from 'lodash-es/dropRight';
-import { generateKeyphrase } from "@cd/actions/createWalletActions";
+import { generateKeyphrase, setNextStep } from "@cd/actions/createWalletActions";
 import { selectCreateWalletTotalKeywords, selectCreateWalletKeyphraseAsMap } from "@cd/selectors/createWallet";
-import useCreateWalletStore from './useCreateWallet';
 import './RecoveryPhrasePage.scss';
 
 const RecoveryPhrasePage = () => {
   const dispatch = useDispatch();
-  const { setNextStep } = useCreateWalletStore();
   const keyPhraseAsMap = useSelector(selectCreateWalletKeyphraseAsMap);
   const TOTAL_KEYWORDS = useSelector(selectCreateWalletTotalKeywords);
   const keyPhraseAsArray = Array.from(keyPhraseAsMap.values());
@@ -18,7 +16,7 @@ const RecoveryPhrasePage = () => {
   const rightKeys = drop(keyPhraseAsArray, TOTAL_KEYWORDS / 2);
   const onClickNextHandler = useCallback(() => {
     setNextStep();
-  }, [setNextStep]);
+  }, []);
 
   useEffect(() => {
     dispatch(generateKeyphrase());
